customized theme
This KIND OF works. The palette property of the Theme object is generated using a createPalette() function, which does not suit my needs. The only downside to not being able to use this is that I'll have to be explicit in my color assignments (e.g. red[500] instead of red.main) and lose out on some other built-in things, like the contrast stuff.