[Game] test.tsx hacky fix
I am casting all of the instances to type `any`.  The alternative solution would be to export the class that I want to cast to separately from my default export. It seems a bit silly to do this exclusively for testing, at a glance. I was further driven to casting to `any` so that I wouldn't have to worry about correctly typing my arguments for the methods I'm calling on the object in question (the Game instance, i.e. the reason for this writeup).