[TileContent] state removal
Got rid of state, to be held in `Game` and passed here as a prop. This also means that `handleClick()` can be removed and all of its logic placed up in `Game` as part of `handleGuess(e)`. I am planning on accessing the guessed word via e.currentTarget.textContent, which is slightly less intuitive but slightly cleaner.