[Game] removed DEFAULT_STATE
It was a bit concerning having to deal with worrying about mutations, so I decided to just write out the stuff two-ish times. My concerns were along the lines of, e.g. https://github.com/Microsoft/TypeScript/issues/10725 , and as a side note, it's a bit annoying that I have to worry about mutating prevState in this.setState.